backend/requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
requests
beautifulsoup4
tqdm
